+++ src/main/java/com/lowdragmc/lowdraglib/gui/widget/FluidStackSelectorWidget.java
@@ -4,7 +4,6 @@ import com.lowdragmc.lowdraglib.gui.texture.ColorBorderTexture;
 import com.lowdragmc.lowdraglib.gui.texture.ResourceBorderTexture;
 import com.lowdragmc.lowdraglib.gui.texture.TextTexture;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
-import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.nbt.TagParser;
 import net.minecraft.resources.ResourceLocation;
@@ -18,7 +17,7 @@ import java.util.Objects;
 import java.util.function.Consumer;
 
 public class FluidStackSelectorWidget extends WidgetGroup {
-    private Consumer<FluidStack> onIFluidStackUpdate;
+    private Consumer<FluidStack> onFluidStackUpdate;
     private final FluidTank handler;
     private final TextFieldWidget fluidField;
     private FluidStack fluid = FluidStack.EMPTY;
@@ -32,18 +31,17 @@ public class FluidStackSelectorWidget extends WidgetGroup {
                 if (fluid == null) {
                     fluid = FluidStack.EMPTY.getFluid();
                 }
-                if (!this.fluid.isFluidEqual(new FluidStack(fluid, 1000))) {
+                if (!FluidStack.isSameFluidSameComponents(this.fluid, new FluidStack(fluid, 1000))) {
                     this.fluid = new FluidStack(fluid, 1000);
                     onUpdate();
                 }
             }
         }).setResourceLocationOnly().setHoverTooltips("ldlib.gui.tips.fluid_selector");
 
-        addWidget(new PhantomFluidWidget(handler = new FluidTank(1000),1, 1)
-                .setIFluidStackUpdater(fluidStack -> {
-                    setIFluidStack(fluidStack);
-                    onUpdate();
-                }).setBackground(new ColorBorderTexture(1, -1)));
+        addWidget(new PhantomFluidWidget(handler = new FluidTank(1000), 0, 1, 1, 18, 18, this::getFluidStack, fluidStack -> {
+            setFluidStack(fluidStack);
+            onUpdate();
+        }).setBackground(new ColorBorderTexture(1, -1)));
         addWidget(fluidField);
         /*
         addWidget(new ButtonWidget(width - 21, 0, 20, 20, null, cd -> {
@@ -68,11 +66,11 @@ public class FluidStackSelectorWidget extends WidgetGroup {
         */
     }
 
-    public FluidStack getIFluidStack() {
+    public FluidStack getFluidStack() {
         return fluid;
     }
 
-    public FluidStackSelectorWidget setIFluidStack(FluidStack fluidStack) {
+    public FluidStackSelectorWidget setFluidStack(FluidStack fluidStack) {
         fluid = Objects.requireNonNullElse(fluidStack, FluidStack.EMPTY).copy();
         if (fluid != FluidStack.EMPTY) {
             fluid.setAmount(1000);
@@ -82,15 +80,15 @@ public class FluidStackSelectorWidget extends WidgetGroup {
         return this;
     }
 
-    public FluidStackSelectorWidget setOnIFluidStackUpdate(Consumer<FluidStack> onIFluidStackUpdate) {
-        this.onIFluidStackUpdate = onIFluidStackUpdate;
+    public FluidStackSelectorWidget setOnFluidStackUpdate(Consumer<FluidStack> onFluidStackUpdate) {
+        this.onFluidStackUpdate = onFluidStackUpdate;
         return this;
     }
 
     private void onUpdate() {
         handler.setFluid(fluid);
-        if (onIFluidStackUpdate != null) {
-            onIFluidStackUpdate.accept(fluid);
+        if (onFluidStackUpdate != null) {
+            onFluidStackUpdate.accept(fluid);
         }
     }
 }